+++ packages/cadl-ranch-dashboard/src/apis.ts
@@ -1,9 +1,11 @@
 import {
   CadlRanchCoverageClient,
-  ResolvedCoverageReport,
-  ScenarioManifest,
   GeneratorMode,
+  ResolvedCoverageReport,
+  ScenarioData,
+  ScenarioManifest,
 } from "@azure-tools/cadl-ranch-coverage-sdk";
+import { TableConfig, TableConfigs } from "./constants.js";
 
 const storageAccountName = "azuresdkcadlranch";
 
@@ -35,7 +37,7 @@ const generatorNames: GeneratorNames[] = [
 export interface CoverageSummary {
   manifest: ScenarioManifest;
   generatorReports: Record<GeneratorNames, ResolvedCoverageReport | undefined>;
-  mode: string;
+  tableConfig: TableConfig;
 }
 
 let client: CadlRanchCoverageClient | undefined;
@@ -57,11 +59,17 @@ export async function getCoverageSummaries(): Promise<CoverageSummary[]> {
     coverageClient.manifest.get(),
     loadReports(coverageClient, generatorNames),
   ]);
-  return GeneratorMode.map((mode) => ({
-    manifest,
-    generatorReports: generatorReports[mode],
-    mode,
-  }));
+  return TableConfigs.map((config) => {
+    const copiedManifest: ScenarioManifest = JSON.parse(JSON.stringify(manifest));
+    copiedManifest.scenarios = manifest.scenarios.filter((scenarioData: ScenarioData) =>
+      config.scenarioFilter ? config.scenarioFilter(scenarioData.name) : () => true,
+    );
+    return {
+      manifest: copiedManifest,
+      generatorReports: generatorReports[config.mode],
+      tableConfig: config,
+    };
+  });
 }
 
 async function loadReports(
@@ -69,22 +77,24 @@ async function loadReports(
   generatorNames: GeneratorNames[],
 ): Promise<{ [mode: string]: Record<GeneratorNames, ResolvedCoverageReport | undefined> }> {
   const results = await Promise.all(
-    GeneratorMode.map(async (mode): Promise<[string, Record<GeneratorNames, ResolvedCoverageReport | undefined>]> => {
-      const items = await Promise.all(
-        generatorNames.map(async (generatorName): Promise<[GeneratorNames, ResolvedCoverageReport | undefined]> => {
-          try {
-            const report = await coverageClient.coverage.getLatestCoverageFor(generatorName, mode);
-            return [generatorName, report];
-          } catch (error) {
-            // eslint-disable-next-line no-console
-            console.error("Error resolving report", error);
+    Object.keys(GeneratorMode).map(
+      async (mode): Promise<[string, Record<GeneratorNames, ResolvedCoverageReport | undefined>]> => {
+        const items = await Promise.all(
+          generatorNames.map(async (generatorName): Promise<[GeneratorNames, ResolvedCoverageReport | undefined]> => {
+            try {
+              const report = await coverageClient.coverage.getLatestCoverageFor(generatorName, mode);
+              return [generatorName, report];
+            } catch (error) {
+              // eslint-disable-next-line no-console
+              console.error("Error resolving report", error);
 
-            return [generatorName, undefined];
-          }
-        }),
-      );
-      return [mode, Object.fromEntries(items) as any];
-    }),
+              return [generatorName, undefined];
+            }
+          }),
+        );
+        return [mode, Object.fromEntries(items) as any];
+      },
+    ),
   );
 
   return results.reduce<{ [mode: string]: Record<GeneratorNames, ResolvedCoverageReport | undefined> }>(