+++ packages/cadl-ranch-specs/http/authentication/api-key/mockapi.ts
@@ -1,4 +1,4 @@
-import { json, MockRequest, passOnSuccess } from "@azure-tools/cadl-ranch-api";
+import { json, passOnCode, passOnSuccess } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
@@ -14,19 +14,15 @@ Scenarios.Authentication_ApiKey_valid = passOnSuccess({
   response: {
     status: 204,
   },
-  handler: (req: MockRequest) => {
-    req.expect.containsHeader("x-ms-api-key", "valid-key");
-    return { status: 204 };
-  },
   kind: "MockApiDefinition",
 });
 
-Scenarios.Authentication_ApiKey_invalid = passOnSuccess({
+Scenarios.Authentication_ApiKey_invalid = passOnCode(403, {
   uri: `/authentication/api-key/invalid`,
   method: `get`,
   request: {
     headers: {
-      "x-ms-api-key": "valid-key",
+      "x-ms-api-key": "invalid-key",
     },
     status: 403,
   },
@@ -35,14 +31,6 @@ Scenarios.Authentication_ApiKey_invalid = passOnSuccess({
     body: json({
       error: "invalid-api-key",
     }),
-  },
-  handler: (req: MockRequest) => {
-    return {
-      status: 403,
-      body: json({
-        error: "invalid-api-key",
-      }),
-    };
   },
   kind: "MockApiDefinition",
 });