+++ packages/cadl-ranch-specs/http/authentication/oauth2/mockapi.ts
@@ -1,4 +1,4 @@
-import { json, MockRequest, passOnSuccess } from "@azure-tools/cadl-ranch-api";
+import { json, passOnSuccess, passOnCode } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
@@ -14,14 +14,10 @@ Scenarios.Authentication_OAuth2_valid = passOnSuccess({
   response: {
     status: 204,
   },
-  handler: (req: MockRequest) => {
-    req.expect.containsHeader("authorization", "Bearer https://security.microsoft.com/.default");
-    return { status: 204 };
-  },
   kind: "MockApiDefinition",
 });
 
-Scenarios.Authentication_OAuth2_invalid = passOnSuccess({
+Scenarios.Authentication_OAuth2_invalid = passOnCode(403, {
   uri: `/authentication/oauth2/invalid`,
   method: "get",
   request: {
@@ -32,14 +28,6 @@ Scenarios.Authentication_OAuth2_invalid = passOnSuccess({
     body: json({
       error: "invalid-grant",
     }),
-  },
-  handler: (req: MockRequest) => {
-    return {
-      status: 403,
-      body: json({
-        error: "invalid-grant",
-      }),
-    };
   },
   kind: "MockApiDefinition",
 });