+++ packages/cadl-ranch-specs/http/azure/client-generator-core/flatten-property/mockapi.ts
@@ -1,6 +1,5 @@
-import { passOnSuccess, json, MockApiDefinition, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, json, MockApiDefinition } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
-import { Mock } from "node:test";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
 function createMockApiDefinitions(route: string, request: any, response: any): MockApiDefinition {
@@ -13,13 +12,6 @@ function createMockApiDefinitions(route: string, request: any, response: any): M
     response: {
       status: 200,
       body: json(response),
-    },
-    handler: (req: MockRequest) => {
-      req.expect.bodyEquals(request);
-      return {
-        status: 200,
-        body: json(response),
-      };
     },
     kind: "MockApiDefinition",
   };