+++ packages/cadl-ranch-specs/http/azure/example/basic/mockapi.ts
@@ -1,6 +1,5 @@
-import { passOnSuccess, json, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, json } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
-import { Mock } from "node:test";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
 
@@ -33,30 +32,6 @@ Scenarios.Client_AzureExampleClient_basicAction = passOnSuccess({
     body: json({
       stringProperty: "text",
     }),
-  },
-  handler: (req: MockRequest) => {
-    req.expect.containsQueryParam("api-version", "2022-12-01-preview");
-    req.expect.containsQueryParam("query-param", "query");
-    req.expect.containsHeader("header-param", "header");
-    const validBody = {
-      stringProperty: "text",
-      modelProperty: {
-        int32Property: 1,
-        float32Property: 1.5,
-        enumProperty: "EnumValue1",
-      },
-      arrayProperty: ["item"],
-      recordProperty: {
-        record: "value",
-      },
-    };
-    req.expect.bodyEquals(validBody);
-    return {
-      status: 200,
-      body: json({
-        stringProperty: "text",
-      }),
-    };
   },
   kind: "MockApiDefinition",
 });