+++ packages/cadl-ranch-specs/http/client/structure/common/service.ts
@@ -1,4 +1,4 @@
-import { MockApiDefinition, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { MockApiDefinition } from "@azure-tools/cadl-ranch-api";
 
 export function createServerTests(uri: string): MockApiDefinition {
   return {
@@ -6,9 +6,6 @@ export function createServerTests(uri: string): MockApiDefinition {
     method: "post",
     request: {},
     response: { status: 204 },
-    handler: (req: MockRequest) => {
-      return { status: 204 };
-    },
     kind: "MockApiDefinition",
   };
 }