+++ packages/cadl-ranch-specs/http/special-words/mockapi.ts
@@ -1,4 +1,4 @@
-import { passOnSuccess, ScenarioMockApi, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
 
@@ -12,11 +12,6 @@ function createOperationsTests(uri: string) {
     request: {},
     response: {
       status: 204,
-    },
-    handler: (req: MockRequest) => {
-      return {
-        status: 204,
-      };
     },
     kind: "MockApiDefinition",
   });
@@ -68,12 +63,6 @@ function createParametersTests(uri: string, data: any, paramName: string) {
     },
     response: {
       status: 204,
-    },
-    handler: (req: MockRequest) => {
-      req.expect.containsQueryParam(paramName, "ok");
-      return {
-        status: 204,
-      };
     },
     kind: "MockApiDefinition",
   });
@@ -334,12 +323,6 @@ function createModelsTests(uri: string) {
     response: {
       status: 204,
     },
-    handler: (req: MockRequest) => {
-      req.expect.bodyEquals({ name: "ok" });
-      return {
-        status: 204,
-      };
-    },
     kind: "MockApiDefinition",
   });
 }
@@ -390,12 +373,6 @@ Scenarios.SpecialWords_ModelProperties_sameAsModel = passOnSuccess({
   },
   response: {
     status: 204,
-  },
-  handler: (req: MockRequest) => {
-    req.expect.bodyEquals({ ["SameAsModel"]: "ok" });
-    return {
-      status: 204,
-    };
   },
   kind: "MockApiDefinition",
 });