+++ packages/cadl-ranch-specs/http/type/union/mockapi.ts
@@ -1,4 +1,4 @@
-import { passOnSuccess, json, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, json } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
@@ -11,9 +11,6 @@ function createGetServerTests(url: string, value: unknown) {
     response: {
       status: 200,
       body: json({ prop: value }),
-    },
-    handler: (req: MockRequest) => {
-      return { status: 200, body: json({ prop: value }) };
     },
     kind: "MockApiDefinition",
   });
@@ -30,10 +27,6 @@ function createPostServerTests(url: string, value: unknown) {
     },
     response: {
       status: 204,
-    },
-    handler: (req: MockRequest) => {
-      req.expect.bodyEquals({ prop: value });
-      return { status: 204 };
     },
     kind: "MockApiDefinition",
   });