+++ packages/cadl-ranch-specs/http/versioning/renamedFrom/mockapi.ts
@@ -1,4 +1,4 @@
-import { passOnSuccess, json, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, json } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
@@ -20,14 +20,6 @@ Scenarios.Versioning_RenamedFrom_newOp = passOnSuccess({
     status: 200,
     body: json({ newProp: "foo", enumProp: "newEnumMember", unionProp: 10 }),
   },
-  handler: (req: MockRequest) => {
-    req.expect.bodyEquals({ newProp: "foo", enumProp: "newEnumMember", unionProp: 10 });
-    req.expect.containsQueryParam("newQuery", "bar");
-    return {
-      status: 200,
-      body: json({ newProp: "foo", enumProp: "newEnumMember", unionProp: 10 }),
-    };
-  },
   kind: "MockApiDefinition",
 });
 
@@ -44,13 +36,6 @@ Scenarios.Versioning_RenamedFrom_NewInterface = passOnSuccess({
   response: {
     status: 200,
     body: json({ newProp: "foo", enumProp: "newEnumMember", unionProp: 10 }),
-  },
-  handler: (req: MockRequest) => {
-    req.expect.bodyEquals({ newProp: "foo", enumProp: "newEnumMember", unionProp: 10 });
-    return {
-      status: 200,
-      body: json({ newProp: "foo", enumProp: "newEnumMember", unionProp: 10 }),
-    };
   },
   kind: "MockApiDefinition",
 });