+++ packages/cadl-ranch-specs/http/versioning/typeChangedFrom/mockapi.ts
@@ -1,4 +1,4 @@
-import { passOnSuccess, json, MockRequest } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, json } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
@@ -18,14 +18,6 @@ Scenarios.Versioning_TypeChangedFrom_test = passOnSuccess({
   response: {
     status: 200,
     body: json({ prop: "foo", changedProp: "bar" }),
-  },
-  handler: (req: MockRequest) => {
-    req.expect.bodyEquals({ prop: "foo", changedProp: "bar" });
-    req.expect.containsQueryParam("param", "baz");
-    return {
-      status: 200,
-      body: json({ prop: "foo", changedProp: "bar" }),
-    };
   },
   kind: "MockApiDefinition",
 });